refactor(react-head): move applyHeadTagsToDOM to its own file
Single-export file per code standards, with its node-environment test
co-located.
