+++ packages/react/head/src/lib/useHead.node.test.tsx
@@ -1,7 +1,7 @@
 // @vitest-environment node
 import { renderToString } from "react-dom/server";
 import { describe, expect, it } from "vitest";
-import useHead, { applyHeadTagsToDOM } from "./useHead.js";
+import useHead from "./useHead.js";
 
 function Probe(): null {
   useHead({
@@ -13,17 +13,6 @@ function Probe(): null {
 }
 
 describe("useHead in non-DOM environments", () => {
-  it("applyHeadTagsToDOM is a no-op without a document", () => {
-    expect(typeof document).toBe("undefined");
-    expect(
-      applyHeadTagsToDOM({
-        title: "Title",
-        meta: [{ name: "description", content: "Desc" }],
-        link: [{ rel: "canonical", href: "https://example.com" }],
-      }),
-    ).toEqual([]);
-  });
-
   it("renders via renderToString without a HeadProvider", () => {
     expect(() => renderToString(<Probe />)).not.toThrow();
   });
